added seperate files per encoder and put retain_graph in if statement
