+++ models/pure_graph_encoder.py
@@ -1,0 +1,27 @@
+from torch import nn
+import torch.nn.functional as F
+from torch_geometric.nn import GCNConv
+
+class PureGraphEncoder(nn.Module):
+    def __init__(self, input_dim, hidden_dim):
+        super(PureGraphEncoder, self).__init__()
+        self.conv1 = GCNConv(input_dim, hidden_dim)
+        self.conv2 = GCNConv(hidden_dim, hidden_dim)
+
+    def forward(self, data, mode):
+        x, edge_index, edge_weight = data.x, data.edge_index, data.edge_attr
+        x = self.conv1(x, edge_index, edge_weight)
+        x = F.relu(x)
+        x = F.dropout(x, training=self.training)
+        x = self.conv2(x, edge_index, edge_weight)
+        if mode == 'train':
+            mask = data.train_mask
+        elif mode == 'val':
+            mask = data.val_mask
+        elif mode == 'test':
+            mask = data.test_mask
+        else:
+            raise ValueError("Mode '%s' is not supported in forward of graph classifier." % mode)
+
+        x = x[mask]
+        return x, data.y[mask]
